Mitra Project
Kael’zak was the last of a bloodline of hunters, living in the depths of the Traan forest. Mundane and forlorn his life had become, after losing his family to a herd of rampaging beasts, he had forsaken his family’s hunting rituals, living on the fruit of the nearby trees. “He who consumes a beast’s food shall be hunted like a beast” his late father would say, but Kael’zak did not feel his ancestral heritage reigns anymore as he believed his bloodline to be doomed, hopelessly waiting his own eventual demise. 
The night had blanketed the Traan forest, as she had done so, quotidian, in all these years of Kael’zak’s solitary life, but his heart was gripped in bitter sadness more than any of those other nights. He could find no remedy to his peculiar sadness and despair, as the morning sun would bring the 10th anniversary of his family’s demise, nonetheless, and so, he wept silently until he fell asleep.
“Kael’zak, help me,” his father screamed. He ran to the hill from the other side of which he could hear his father’s cries. From the hilltop he saw his father under a tree, being beaten by three strange humanoid creatures. But, as he tried to run toward his father, the forest was ruptured under his feet, devouring him into darkness and suffocation. 
Kael’zak stood up, feeling the warmth of the sun on his body, the nightmare already fading away from his mind. His blood boiling with anger and his ears still echoing the cries of his father, he picked up his long forgotten hunting bow, hurrying to the heart of the Traan forest, where he had seen his father in the nightmare. He reached the same hill, and he, unbelievably, saw a frail female figure collapsed under the same tree, her beautiful face glowing with a white heavenly light. 
“Who are you? Where am I?” she asked, as she came to, after two days of being nursed by the retired hunter.
“I am Kael’zak Silvershade, the last descendant of Iraj, the great hunter of the Traan forest,” he said, “Who are you, suddenly appearing in the forest and calling me in my dreams?”
“Someone heard my cries then,” she said as she seemed to be recovering her memories, “I am Mitra Moonface, messenger of the Lord of the Realms.”
“Lord of the Realms?” he asked.
Mitra paused for a moment and then moved her hands as if she was praying to the goddess of the moon. For a moment, Kael’zak could see her face glowing brighter than ever and her red hair floating in the air.
“Look at your left hand,” she said, stopping her prayer, and as Kael’zak was looking at the strange glittering symbols on the palm of his hand, she continued: “This is your realm’s charm.”
“Now listen carefully, Kael’zak,” she said, “the world consists of numerous realms, each with its own charm. Each being has her own realm’s charm carved into their soul, preventing them from leaving their realm. Only we of the Moonface bloodline can freely move between the realms, and we have been blessed with primordial and eternal knowledge of the realms by the Lord of the Realms to become his faithful messengers. I need you to take this message from me and deliver it to a certain realm.” She concluded, showing Kael’zak a tiny scroll.

“But,” he protested, “you said you are the only bloodline able to move between the realms, why are you asking me to deliver your message?”
“I have been rendered powerless, as you can see, and there are ways to move between the realms even for you. As I said, every realm has a unique charm, if you change your charm to that of another realm, you will vanish from this realm and you will be born anew in the other.”
"But how can I change my charm?” asked the hunter.
“There are magi of different powers that can use their incantations to change the charm carved into your soul. Each incantation can either add a symbol, remove a symbol or change a symbol to another in your charm. But, you need to be careful, if any magus uses their incantation on your charm, only magi of greater power than the previous magus can change your charm.”
“But where are these magi? Why would they help me?”
“The magi have always been here. You have met them, but you have not seen them. Now that I have enchanted your eyes with the power of Gyo, you will be able to see them, just as you are able to see your charm in your hand. The magi will ask you for a quantity of gems equal to their power, and if you provide them with the gems they will use their incantation of your charm.”
“And how can I find these magi?”“In each realm in which you are born or reborn, you will meet a magus on each of your birthdays.”Mitra sighed and continued: “So remember Kael’zak, on each realm you are going to meet a magus every
year. Each magus can use their incantation to either add a symbol, remove a symbol or change a symbol in your charm. If a magus changes your charm in a realm, only magi of greater power will be able to change it while you are in that same realm. Each magus will ask for a number of gems equal to their power if you want them to change your charm.”
She then said: “Now I will give you the charm for all of the realms, and I will also tell you the power of the magi that you would meet in each realm in the order that they would meet you, if you were reborn in those realms. I will also tell you the charm for the realm to which you need to deliver the message. You have to find the way with the least number of incantations needed to reach that realm, and also the way back from there to this realm. For each way, you have to tell me the least number of incantations needed, and the number of gems you will need throughout the way.”
And so, Kael’zak the hunter, started to learn the arts of data structures and algorithms in order to give the correct answer to Mitra the messenger. 

Input
On the first line you can read the number of realms,N , on the next 3×N lines you can read, on each first line the charm for the realm, on each second line the number of magi which Kael’zak would meet in that realm and on each third line the power of the aforementioned magi in the order that they would meet Kael’zak in the realm. On the following two lines you can read, respectively, the charm for Kael’zak’s realm and the charm for the destination realm.

Output
On the first line you should print two numbers, the minimum number of incantations needed and the number of gems needed for going from Kael’zak’s realm to the destination realm. There may be multiple correct answers for the number of gems. On the second line you should print the same information for coming back from the destination realm to Kael’zak’s realm. If it’s impossible to reach the destination from Kael’zak’s realm, you should print “IMPOSSIBLE” in the first line. If it’s impossible to come back from destination to the Kael’zak’s realm, you should print “IMPOSSIBLE” in the second line.